Upgrade tests: common imports and constants
